linked pages, done. More Sass did not happen.
Using the &__ feature with Sass wasn't really working with me on this project. I tried to focus on writing clear BEM-names to understand the structure better in hopes that I could Sass it later. But it didn't work so well. Therefor there is only a little bit of sass on the projects.scss file.